# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/LoginResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Malformed login payload.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Login rate limit exceeded (5 / 15 min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Authenticate; sets the refresh cookie and returns an access token.",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/RefreshResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired, or revoked refresh token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Refresh rate limit exceeded (30 / hour).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "summary": "Rotate the refresh token and issue a new access token.",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/v1/admin/categories": {
      "get": {
        "operationId": "CategoriesAdminController_list_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminCategoryEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List categories with full translation maps.",
        "tags": [
          "taxonomy"
        ]
      },
      "post": {
        "operationId": "CategoriesAdminController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminCategoryEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or slug already in use.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a category with its translations.",
        "tags": [
          "taxonomy"
        ]
      }
    },
    "/api/v1/admin/categories/{id}": {
      "patch": {
        "operationId": "CategoriesAdminController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "data"
                  ],
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AdminCategoryEntity"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or slug already in use.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Upsert category translations.",
        "tags": [
          "taxonomy"
        ]
      },
      "delete": {
        "operationId": "CategoriesAdminController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted."
          },
          "401": {
            "description": "Missing or invalid access token.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing the required permission.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "409": {
            "description": "Category is still referenced by articles.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          },
          "429": {
            "description": "Admin rate limit exceeded (300 / min).",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetailsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a category (409 if articles still reference it).",
        "tags": [
          "taxonomy"
        ]
      }
    },
...
  },
  "info": {
    "title": "eslammuatamed API",
    "description": "REST contract for the eslammuatamed platform. Public reads are unauthenticated and locale-resolved; the /admin surface requires a bearer access token (doc 10, doc 19).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "FieldErrorDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "example": "translations[0].slug",
            "description": "Dotted path to the offending field."
          },
          "message": {
            "type": "string",
            "example": "Slug already exists for locale 'en'."
          }
        },
        "required": [
          "field",
          "message"
        ]
      },
      "ProblemDetailsDto": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "example": "/problems/validation",
            "description": "A URI reference identifying the problem type."
          },
          "title": {
            "type": "string",
            "example": "Validation failed"
          },
          "status": {
            "type": "number",
            "example": 422
          },
          "detail": {
            "type": "string",
            "example": "2 fields failed validation."
          },
          "instance": {
            "type": "string",
            "example": "/api/v1/admin/articles"
          },
          "errors": {
            "description": "Present on 422 responses only.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldErrorDto"
            }
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ]
      },
...
      "AuthUserRoleEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "018f9d3c-1a2b-7c3d-8e4f-5a6b7c8d9e0f",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "OWNER"
          }
        },
        "required": [
          "id",
          "name"
        ]
      },
      "AuthUserEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "018f9d3c-1a2b-7c3d-8e4f-5a6b7c8d9e0f",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "example": "owner@example.com",
            "format": "email"
          },
          "role": {
            "$ref": "#/components/schemas/AuthUserRoleEntity"
          }
        },
        "required": [
          "id",
          "email",
          "role"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "15-minute access token; the refresh token is set as an httpOnly cookie."
          },
          "user": {
            "$ref": "#/components/schemas/AuthUserEntity"
          }
        },
        "required": [
          "accessToken",
          "user"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "owner@example.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "change-me-minimum-12-chars",
            "minLength": 1
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "accessToken"
        ]
      },
...
      "CategoryTranslationEntity": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Engineering"
          },
          "slug": {
            "type": "string",
            "example": "engineering"
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "slug",
          "description"
        ]
      },
      "AdminCategoryEntity": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "translations": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/CategoryTranslationEntity"
            },
            "description": "Translation map keyed by locale code."
          }
        },
        "required": [
          "id",
          "translations"
        ]
      },
      "CategoryTranslationDto": {
        "type": "object",
        "properties": {
          "locale": {
            "type": "string",
            "example": "en",
            "description": "Two-letter locale; must be enabled."
          },
          "name": {
            "type": "string",
            "example": "Engineering"
          },
          "slug": {
            "type": "string",
            "example": "engineering"
          },
          "description": {
            "type": "string",
            "example": "Systems, architecture, and craft."
          }
        },
        "required": [
          "locale",
          "name",
          "slug"
        ]
      },
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "translations": {
            "description": "At least one locale translation.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryTranslationDto"
            }
          }
        },
        "required": [
          "translations"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "translations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryTranslationDto"
            }
          }
        }
      },
...
    }
  }
}